+++ .devcontainer/.internal_file_registry_service.yaml
