+++ .devcontainer/dev_launcher
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 # adapt to package name
-my-microservice --config /workspace/.devcontainer/dev_config.yaml
+internal-file-registry-service --config /workspace/.devcontainer/dev_config.yaml
