+++ Dockerfile
@@ -28,4 +28,4 @@ USER appuser
 ENV PYTHONUNBUFFERED=1
 
 # Please adapt to package name:
-ENTRYPOINT ["my-microservice"]
+ENTRYPOINT ["internal-file-registry-service"]
