+++ README.md
@@ -10,14 +10,14 @@ from the GHGA Dev Team.
 
 ## Naming Conventions
 The github repository contains only lowercase letters, numbers, and hyphens "-",
-e.g.: `my-microservice`
+e.g.: `internal-file-registry-service`
 
 The python package (and thus the source repository) contains underscores "_"
-instead of hyphens, e.g.: `my_microservice`
+instead of hyphens, e.g.: `internal_file_registry_service`
 
 The command-line script that is used to run the service, the docker repository
 (published to docker hub), and the helm chart (not part of this repository) use the
-same pattern as the repository name, e.g.: `my-microservice`
+same pattern as the repository name, e.g.: `internal-file-registry-service`
 ## Adapt to your service
 This is just a template and needs some adaption to your specific use case.
 
@@ -26,14 +26,14 @@ that need modification. Once the adaptions are in place, please remove these #
 comments.
 
 The following should serve as a template for the final repo's README,
-please adapt it accordingly (e.g. replace all occurences of `my-microservice` or `my_microservice` with the final package name and don't forget to adapt the links):
+please adapt it accordingly (e.g. replace all occurences of `internal-file-registry-service` or `internal_file_registry_service` with the final package name and don't forget to adapt the links):
 
 ---
 
 **\# please adapt the links of following badges:**
 ![tests](https://github.com/ghga-de/microservice-repository-template/actions/workflows/unit_and_int_tests.yaml/badge.svg)
 [![codecov](https://codecov.io/gh/ghga-de/microservice-repository-template/branch/main/graph/badge.svg?token=GYH99Y71CK)](https://codecov.io/gh/ghga-de/microservice-repository-template)
-# My-Microservice - A Short Description
+# Internal-File-Registry-Service - A Short Description
 
 A longer description explaining the use case of this service.
 
@@ -45,17 +45,17 @@ An extensive documentation can be found [here](...) (coming soon).
 ### Installation
 We recommend using the provided Docker container.
 
-A pre-build version is available at [docker hub](https://hub.docker.com/repository/docker/ghga/my-microservice):
+A pre-build version is available at [docker hub](https://hub.docker.com/repository/docker/ghga/internal-file-registry-service):
 ```bash
 # Please feel free to choose the version as needed:
-docker pull ghga/my-microservice:<version>
+docker pull ghga/internal-file-registry-service:<version>
 ```
 
 Or you can build the container yourself from the [`./Dockerfile`](./Dockerfile):
 ```bash
 # Execute in the repo's root dir:
 # (Please feel free to adapt the name/tag.)
-docker build -t ghga/my-microservice:<version> .
+docker build -t ghga/internal-file-registry-service:<version> .
 ```
 
 For production-ready deployment, we recommend using Kubernetes, however,
@@ -63,7 +63,7 @@ for simple use cases, you could execute the service using docker
 on a single server:
 ```bash
 # The entrypoint is preconfigured:
-docker run -p 8080:8080 ghga/my-microservice:<version>
+docker run -p 8080:8080 ghga/internal-file-registry-service:<version>
 ```
 
 If you prefer not to use containers, you may install the service from source:
@@ -72,14 +72,14 @@ If you prefer not to use containers, you may install the service from source:
 pip install .
 
 # to run the service:
-my-microservice
+internal-file-registry-service
 ```
 
 ### Configuration:
 The [`./example-config.yaml`](./example-config.yaml) gives an overview of the available configuration options.
-Please adapt it, rename it to `.my_microservice.yaml`, and place it to one of the following locations:
-- in the current working directory were you are execute the service (on unix: `./.my_microservice.yaml`)
-- in your home directory (on unix: `~/.my_microservice.yaml`)
+Please adapt it, rename it to `.internal_file_registry_service.yaml`, and place it to one of the following locations:
+- in the current working directory were you are execute the service (on unix: `./.internal_file_registry_service.yaml`)
+- in your home directory (on unix: `~/.internal_file_registry_service.yaml`)
 
 The config yaml will be automatically parsed by the service.
