+++ db_migration/alembic/env.py
@@ -18,7 +18,7 @@ config.set_main_option("sqlalchemy.url", os.getenv("DB_URL"))
 # Please adapt to package name
 # add your model's MetaData object here
 # for 'autogenerate' support
-from my_microservice.dao.db_models import Base
+from internal_file_registry_service.dao.db_models import Base
 
 target_metadata = Base.metadata
 # target_metadata = None
