+++ setup.cfg
@@ -15,9 +15,9 @@
 
 [metadata]
 # Please adapt to package name:
-name = my_microservice
-version = attr: my_microservice.__version__
-description = My-Microservice - a short description
+name = internal_file_registry_service
+version = attr: internal_file_registry_service.__version__
+description = Internal-File-Registry-Service - a short description
 url = https://github.com/ghga-de/microservice-repository-template
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8
@@ -63,7 +63,7 @@ python_requires = >= 3.9
 [options.entry_points]
 # Please adapt to package name:
 console_scripts =
-    my-microservice = my_microservice.__main__:run
+    internal-file-registry-service = internal_file_registry_service.__main__:run
 
 [options.extras_require]
 # Please adapt:
